Fixing BackhgroundImage Bug and various other improvements.
Reorganized the code for better readability and clarity.
Used private access control for the state variables.
Combined the Tab views into one TabView struct to reduce redundancy.
Utilized a switch statement in the SideMenuView closure for easier handling of tab selection.
Renamed variables for clarity and consistency.
Applied a default case for unexpected tab titles in the switch statement